refactor: extract workflow types generation into utility function
Extract duplicated workflow types generation logic from workflow and workflow-types commands into a shared utility function to improve code maintainability and reduce duplication.
