+++ packages/cli/src/functions/wirings/workflow/pikku-command-workflow.ts
@@ -6,12 +6,11 @@ import {
   serializeWorkflowMeta,
   serializeWorkflowMetaTS,
 } from './serialize-workflow-meta.js'
-import { serializeWorkflowTypes } from './serialize-workflow-types.js'
 import { serializeWorkflowMap } from './serialize-workflow-map.js'
 import { serializeWorkflowWorkers } from './serialize-workflow-workers.js'
 import { serializeWorkflowWirings } from './serialize-workflow-wirings.js'
 import { getFileImportRelativePath } from '../../../utils/file-import-path.js'
-import { checkRequiredTypes } from '../../../utils/check-required-types.js'
+import { generateWorkflowTypes } from './utils/generate-workflow-types.js'
 import { join } from 'path'
 
 export const pikkuWorkflow: any = pikkuSessionlessFunc<
@@ -34,22 +33,6 @@ export const pikkuWorkflow: any = pikkuSessionlessFunc<
     } = config
     const { workflows, functions: functionState } = visitState
     const { typesMap } = functionState
-
-    // Check for required types
-    checkRequiredTypes(visitState.filesAndMethodsErrors, {
-      userSessionType: true,
-      sessionServiceType: true,
-      singletonServicesType: true,
-      singletonServicesFactory: false,
-      sessionServicesFactory: false,
-    })
-
-    const { userSessionType, singletonServicesType } =
-      visitState.filesAndMethods
-
-    if (!userSessionType || !singletonServicesType) {
-      throw new Error('Required types not found')
-    }
 
     // Validate that workflowService service is configured if workflows are defined
     const hasWorkflows = Object.keys(workflows.meta).length > 0
@@ -104,42 +87,14 @@ export const pikkuWorkflow: any = pikkuSessionlessFunc<
     )
 
     // Write workflow types
-    const functionTypesImportPath = getFileImportRelativePath(
+    const workflowTypes = generateWorkflowTypes(visitState, {
       workflowTypesFile,
       functionTypesFile,
-      packageMappings
-    )
-
-    const userSessionImportPath = getFileImportRelativePath(
-      workflowTypesFile,
-      userSessionType.typePath,
-      packageMappings
-    )
-
-    const singletonServicesImportPath = getFileImportRelativePath(
-      workflowTypesFile,
-      singletonServicesType.typePath,
-      packageMappings
-    )
-
-    const rpcMapImportPath = getFileImportRelativePath(
-      workflowTypesFile,
       rpcInternalMapDeclarationFile,
-      packageMappings
-    )
+      packageMappings,
+    })
 
-    await writeFileInDir(
-      logger,
-      workflowTypesFile,
-      serializeWorkflowTypes(
-        functionTypesImportPath,
-        userSessionImportPath,
-        userSessionType.type,
-        singletonServicesImportPath,
-        singletonServicesType.type,
-        rpcMapImportPath
-      )
-    )
+    await writeFileInDir(logger, workflowTypesFile, workflowTypes)
 
     // Write workflow map (type-safe client API)
     await writeFileInDir(
